1-10 uploadToAppStore
更新店铺优惠券功能，打包上线版

+++ XuXin/Models/SelfModel/FaceToFaceOrderModel.h
@@ -42,7 +42,7 @@
  14    redpacketmoeny    红包抵扣金额（未使用的情况下为0，使用了就是金额大小）    是    [string]
  */
 
-@property (nonatomic, assign) NSInteger couponmoeny;
-@property (nonatomic, assign) NSInteger redpacketmoeny;
+@property (nonatomic, assign) CGFloat couponmoeny;
+@property (nonatomic, assign) CGFloat redpacketmoeny;
 
 @end
